added test for FHHttpClient
